app/backend/django/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.7.2
attrs==24.2.0
Authlib==1.3.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
conllu==6.0.0
cryptography==44.0.0
cymem==2.0.10
distro==1.9.0
Django==4.1.13
django-cors-headers==4.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.4.2
en_core_sci_sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz#sha256=8588289f761488021e096e8a7ebbdad52acc0e74685e602c5d0d0b7a5fa546a3
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.4
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
networkx==3.4.2
nmslib-metabrainz==2.1.3
numpy==1.26.4
openai==1.57.1
packaging==24.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.0.0
preshed==3.0.9
propcache==0.2.1
psutil==6.1.0
psycopg2-binary==2.9.10
pybind11==2.13.6
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.6.0
pysbd==0.3.4
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rest-framework-simplejwt==0.0.2
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
scispacy==0.5.5
sentence-transformers==3.3.1
shellingham==1.5.4
smart-open==6.4.0
sniffio==1.3.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.2.4
srsly==2.4.8
sympy==1.13.1
thinc==8.1.12
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typer==0.7.0
typing_extensions==4.12.2
urllib3==2.0.7
validators==0.34.0
wasabi==0.10.1
weasel==0.4.1
weaviate-client==3.26.0
wrapt==1.17.0
yarl==1.18.3